services/auth/requirements.txt
aiohttp==3.9.5
aiosqlite==0.20.0
alembic==1.13.2
async-fastapi-jwt-auth==0.6.5
asyncpg==0.29.0
cryptography==42.0.7
databases==0.9.0
elasticsearch[async]==8.13.0
fastapi==0.109.0
orjson==3.10.2
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.10.1
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytest==8.2.0
pytest-asyncio==0.23.7
redis==5.0.4
SQLAlchemy==2.0.30
uvicorn==0.29.0
uvloop==0.19.0 ; sys_platform != "win32" and implementation_name == "cpython"
Werkzeug==3.0.3
email-validator==2.2.0
typer==0.12.3
fastapi-limiter==0.1.6
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
authlib==1.3.1
itsdangerous==2.2.0